feat(payment): add payment callback pages (#30)
- Create payment success, error, and cancel pages
- Add payment demo page for testing SDK flow
- Implement content components for each callback type
- Add API endpoints for payment status handling

Refs: #30
